backend/requirements.txt
# Core Django
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-spectacular==0.27.2

# Database (SQLite for development)
# psycopg2-binary==2.9.9  # Commented out for SQLite

# Authentication & Communication
twilio==9.3.3

# Google Maps integration
googlemaps==4.10.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.1

# Development tools
black==24.8.0
flake8==7.1.1
isort==5.13.2
